pyproject.toml
[project]
name = "linemessagingmicroservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.8",
    "line-bot-sdk>=3.16.0",
    "uvicorn>=0.34.0",
]
